requirements.txt
ta
websocket
python-binance
apprise
configparser
pandas